requirements.txt
Django~=5.0.6
python-decouple==3.8
django-taggit==5.0.1